dithering/requirements.txt
Pillow==8.3.2
numpy==1.21.2
matplotlib==3.4.3
imageio==2.14.0
scikit-image==0.18.3
scikit-learn
opencv-python==4.5.3.56
